Split common module into separate files
