+++ cli/src/structures/common/mod.rs
@@ -1,0 +1,9 @@
+mod config;
+mod file;
+mod generator;
+mod package;
+
+pub use config::*;
+pub use file::*;
+pub use generator::*;
+pub use package::*;
